pyproject.toml
[tool.poetry]
name = "artemis-crm-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.136.0"
uvicorn = "^0.44.0"
sqlalchemy = "^2.0.49"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.2"
passlib = "^1.7.4"
bcrypt = "^3.2.2"
python-jose = "^3.5.0"
alembic = "^1.18.4"
email-validator = "^1.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.25.0"
faker = "^20.0.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"